Make text in emoji items secondary color and the snackbar duration short again

+++ features/roomlist/impl/src/main/kotlin/io/element/android/features/roomlist/impl/RoomListView.kt
@@ -135,7 +135,7 @@ fun RoomListContent(
         if (state.presentVerificationSuccessfulMessage) {
             snackbarHostState.showSnackbar(
                 message = verificationCompleteMessage,
-                duration = SnackbarDuration.Long,
+                duration = SnackbarDuration.Short,
             )
             state.eventSink(RoomListEvents.ClearSuccessfulVerificationMessage)
         }
